Update README.md
A snake game is an arcade maze game which has been developed by Gremlin Industries and published by Sega in October 1976. It is considered to be a skillful game and has been popularized among people for generations. The snake in the Snake game is controlled using the four direction buttons relative to the direction it is headed in. The player’s objective in the game is to achieve maximum points as possible by collecting food or fruits. The player loses once the snake hits the wall or hits itself.

For the python beginners, those who are interested in making something easier in your domain can definitely try this out and the module Turtle was made exactly for this purpose for the beginners to try out and can also submit as a part of the project. This program will be done in Python 3.

So, we will be creating a Python-based-game using the following modules:

Turtle: It is a pre-installed python library that enables users to create shapes and pictures by providing them with a virtual canvas.
Time: This function is used to count the number of seconds elapsed since the epoch.
Random: This function is used to generate random numbers in Python by using random module.